@@ -7,13 +7,17 @@
     "build:lib": "babel ./src -d lib",
     "build:umd": "webpack src/index_umd.js dist/GoogleMapReact.js --config webpack.config.dev.js",
     "build:umd:min": "webpack src/index_umd.js dist/GoogleMapReact.min.js --config webpack.config.prod.js",
-    "build": "npm run build:lib && npm run build:umd && npm run build:umd:min",
+    "build": "yarn run build:lib && yarn run build:umd && yarn run build:umd:min",
     "clean": "rimraf lib dist",
-    "prepublish": "npm run clean && npm run build",
-    "lint": "eslint src develop",
+    "prepublish": "yarn run clean && yarn run build",
+    "format": "prettier --trailing-comma es5 --single-quote --write 'src/**/*.js' 'src/**/*.js'",
+    "format:dev": "prettier --trailing-comma es5 --single-quote --write 'develop/**/*.js' 'develop/**/*.js'",
+    "lint": "eslint src",
+    "lint:dev": "eslint develop",
     "test": "NODE_ENV=eee mocha --compilers js:babel-register --recursive --require babel-polyfill",
     "test:watch": "NODE_ENV=eee mocha --compilers js:babel-register --recursive  --require babel-polyfill --watch",
-    "start": "kotatsu serve --port 4000 --config ./develop/config/index.babel.js --presets es2015,stage-0,react,react-hmre ./develop/Main.js"
+    "start": "kotatsu serve --port 4000 --config ./develop/config/index.babel.js --presets es2015,stage-0,react,react-hmre ./develop/Main.js",
+    "precommit": "lint-staged"
   },
   "repository": {
     "type": "git",
@@ -38,6 +42,7 @@
   },
   "homepage": "https://github.com/istarkov/google-map-react#readme",
   "peerDependencies": {
+    "prop-types": "^15.5.6",
     "react": "^0.14.0 || ^15.0.0"
   },
   "dependencies": {
@@ -60,20 +65,27 @@
     "babel-preset-stage-0": "^6.5.0",
     "babel-register": "^6.14.0",
     "css-loader": "^0.23.1",
-    "eslint": "^2.12.0",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-import": "^1.8.1",
-    "eslint-plugin-jsx-a11y": "^1.4.2",
-    "eslint-plugin-react": "^5.1.1",
+    "eslint": "^3.17.1",
+    "eslint-config-airbnb": "^14.0.0",
+    "eslint-config-prettier": "^1.5.0",
+    "eslint-plugin-babel": "^4.1.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-prettier": "^2.0.1",
+    "eslint-plugin-react": "^6.8.0",
     "expect": "^1.11.1",
     "file-loader": "^0.8.5",
+    "husky": "^0.13.3",
     "jsdom": "^6.5.1",
     "kotatsu": "^0.14.0",
+    "lint-staged": "^3.4.0",
     "lodash": "^4.13.1",
     "mocha": "^2.3.3",
     "node-sass": "^3.7.0",
     "normalize.css": "^4.1.1",
     "postcss-loader": "^0.9.1",
+    "prettier": "^0.22.0",
+    "prop-types": "^15.5.6",
     "react": "^15.1.0",
     "react-addons-test-utils": "^15.1.0",
     "react-dom": "^15.1.0",
@@ -87,5 +99,11 @@
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
     "webpack": "^1.12.2"
+  },
+  "lint-staged": {
+    "*.js": [
+      "prettier --trailing-comma es5 --single-quote --write",
+      "git add"
+    ]
   }
 }
