@@ -8,25 +8,32 @@
 import React from 'react';
 import compose from 'recompose/compose';
 import defaultProps from 'recompose/defaultProps';
-import withStateSelector from './utils/withStateSelector';
 import withHandlers from 'recompose/withHandlers';
 import withState from 'recompose/withState';
 import withPropsOnChange from 'recompose/withPropsOnChange';
 import withProps from 'recompose/withProps';
+import { createSelector } from 'reselect';
+import withStateSelector from './utils/withStateSelector';
 import ptInBounds from './utils/ptInBounds';
 import GoogleMapReact from '../src';
 // import SimpleMarker from './markers/SimpleMarker';
 import ReactiveMarker from './markers/ReactiveMarker';
-import { createSelector } from 'reselect';
 import { susolvkaCoords, generateMarkers } from './data/fakeData';
 import props2Stream from './utils/props2Stream';
 
-export const gMap = ({
-  style, hoverDistance, options,
-  mapParams: { center, zoom },
-  onChange, onChildMouseEnter, onChildMouseLeave,
-  markers, draggable,
-}) => (
+export const gMap = (
+  {
+    style,
+    hoverDistance,
+    options,
+    mapParams: { center, zoom },
+    onChange,
+    onChildMouseEnter,
+    onChildMouseLeave,
+    markers,
+    draggable,
+  }
+) => (
   <GoogleMapReact
     style={style}
     options={options}
@@ -58,56 +65,52 @@ export const gMapHOC = compose(
       flex: 1,
     },
   }),
-
   // withState so you could change markers if you want
-  withStateSelector(
-    'markers',
-    'setMarkers',
-    () => createSelector(
+  withStateSelector('markers', 'setMarkers', () =>
+    createSelector(
       ({ route: { markersCount = 20 } }) => markersCount,
-      (markersCount) => generateMarkers(markersCount)
-    )
-  ),
+      markersCount => generateMarkers(markersCount)
+    )),
   withState('hoveredMarkerId', 'setHoveredMarkerId', -1),
   withState('mapParams', 'setMapParams', { center: susolvkaCoords, zoom: 6 }),
   // describe events
   withHandlers({
-    onChange: ({ setMapParams }) => ({ center, zoom, bounds }) => {
-      setMapParams({ center, zoom, bounds });
-    },
-    onChildMouseEnter: ({ setHoveredMarkerId }) => (hoverKey, { id }) => {
-      setHoveredMarkerId(id);
-    },
-    onChildMouseLeave: ({ setHoveredMarkerId }) => () => {
-      setHoveredMarkerId(-1);
-    },
+    onChange: ({ setMapParams }) =>
+      ({ center, zoom, bounds }) => {
+        setMapParams({ center, zoom, bounds });
+      },
+    onChildMouseEnter: ({ setHoveredMarkerId }) =>
+      (hoverKey, { id }) => {
+        setHoveredMarkerId(id);
+      },
+    onChildMouseLeave: ({ setHoveredMarkerId }) =>
+      () => {
+        setHoveredMarkerId(-1);
+      },
   }),
-  withPropsOnChange(
-    ['markers', 'mapParams'],
-    ({ markers, mapParams: { bounds } }) => ({
-      markers: bounds
-        ? markers.filter(m => ptInBounds(bounds, m))
-        : [],
-    })
-  ),
+  withPropsOnChange(['markers', 'mapParams'], ({
+    markers,
+    mapParams: { bounds },
+  }) => ({
+    markers: bounds ? markers.filter(m => ptInBounds(bounds, m)) : [],
+  })),
   withProps(({ hoveredMarkerId }) => ({
     draggable: hoveredMarkerId === -1,
   })),
   props2Stream('hoveredMarkerId'),
-  withPropsOnChange(
-    ['markers', 'hoveredMarkerId$'],
-    ({ markers, hoveredMarkerId$ }) => ({
-      markers: markers
-        .map(({ ...markerProps, id }) => (
-          <ReactiveMarker
-            key={id}
-            id={id}
-            hoveredMarkerId$={hoveredMarkerId$}
-            {...markerProps}
-          />
-        )),
-    })
-  )
+  withPropsOnChange(['markers', 'hoveredMarkerId$'], ({
+    markers,
+    hoveredMarkerId$,
+  }) => ({
+    markers: markers.map(({ ...markerProps, id }) => (
+      <ReactiveMarker
+        key={id}
+        id={id}
+        hoveredMarkerId$={hoveredMarkerId$}
+        {...markerProps}
+      />
+    )),
+  }))
 );
 
 export default gMapHOC(gMap);
