@@ -14,12 +14,19 @@ import SimpleMarker from './markers/SimpleMarker';
 import { createSelector } from 'reselect';
 import { susolvkaCoords, generateMarkers } from './data/fakeData';
 
-export const gMapResizable = ({
-  style, hoverDistance, options,
-  mapParams: { center, zoom },
-  onChange, onChildMouseEnter, onChildMouseLeave,
-  markers, draggable, // hoveredMarkerId,
-}) => (
+export const gMapResizable = (
+  {
+    style,
+    hoverDistance,
+    options,
+    mapParams: { center, zoom },
+    onChange,
+    onChildMouseEnter,
+    onChildMouseLeave,
+    markers,
+    draggable, // hoveredMarkerId,
+  }
+) => (
   <GoogleMapReact
     draggable={draggable}
     style={style}
@@ -31,11 +38,9 @@ export const gMapResizable = ({
     onChildMouseEnter={onChildMouseEnter}
     onChildMouseLeave={onChildMouseLeave}
     resetBoundsOnResize={true}
-    apiKey={"AIzaSyC-BebC7ChnHPzxQm7DAHYFMCqR5H3Jlps"}
+    apiKey={'AIzaSyC-BebC7ChnHPzxQm7DAHYFMCqR5H3Jlps'}
   >
-    {
-      markers
-    }
+    {markers}
   </GoogleMapReact>
 );
 
@@ -54,57 +59,44 @@ export const gMapHOC = compose(
       flex: 1,
     },
   }),
-  withContext(
-    { hello: PropTypes.string },
-    () => ({ hello: 'world' })
-  ),
+  withContext({ hello: PropTypes.string }, () => ({ hello: 'world' })),
   // withState so you could change markers if you want
-  withStateSelector(
-    'markers',
-    'setMarkers',
-    () => createSelector(
+  withStateSelector('markers', 'setMarkers', () =>
+    createSelector(
       ({ route: { markersCount = 20 } }) => markersCount,
-      (markersCount) => generateMarkers(markersCount)
-    )
-  ),
+      markersCount => generateMarkers(markersCount)
+    )),
   withState('hoveredMarkerId', 'setHoveredMarkerId', -1),
   withState('mapParams', 'setMapParams', { center: susolvkaCoords, zoom: 6 }),
   // describe events
   withHandlers({
-    onChange: ({ setMapParams }) => ({ center, zoom, bounds }) => {
-      setMapParams({ center, zoom, bounds });
-    },
-    onChildMouseEnter: ({ setHoveredMarkerId }) => (hoverKey, { id }) => {
-      setHoveredMarkerId(id);
-    },
-    onChildMouseLeave: ({ setHoveredMarkerId }) => () => {
-      setHoveredMarkerId(-1);
-    },
+    onChange: ({ setMapParams }) =>
+      ({ center, zoom, bounds }) => {
+        setMapParams({ center, zoom, bounds });
+      },
+    onChildMouseEnter: ({ setHoveredMarkerId }) =>
+      (hoverKey, { id }) => {
+        setHoveredMarkerId(id);
+      },
+    onChildMouseLeave: ({ setHoveredMarkerId }) =>
+      () => {
+        setHoveredMarkerId(-1);
+      },
   }),
-  withPropsOnChange(
-    ['markers', 'mapParams'],
-    ({ markers, mapParams: { bounds } }) => ({
-      markers: bounds
-        ? markers.filter(m => ptInBounds(bounds, m))
-        : [],
-    })
-  ),
+  withPropsOnChange(['markers', 'mapParams'], ({
+    markers,
+    mapParams: { bounds },
+  }) => ({
+    markers: bounds ? markers.filter(m => ptInBounds(bounds, m)) : [],
+  })),
   withProps(({ hoveredMarkerId }) => ({
     draggable: hoveredMarkerId === -1,
   })),
-  withPropsOnChange(
-    ['markers'],
-    ({ markers }) => ({
-      markers: markers
-        .map(({ ...markerProps, id }) => (
-          <SimpleMarker
-            key={id}
-            id={id}
-            {...markerProps}
-          />
-        )),
-    })
-  )
+  withPropsOnChange(['markers'], ({ markers }) => ({
+    markers: markers.map(({ ...markerProps, id }) => (
+      <SimpleMarker key={id} id={id} {...markerProps} />
+    )),
+  }))
 );
 
 export default gMapHOC(gMapResizable);
