@@ -1,14 +1,14 @@
-// file: main.jsx
+/* eslint-disable import/no-named-as-default */
+
 import React from 'react';
 import { render } from 'react-dom';
 import { Router, Route, IndexRoute, browserHistory } from 'react-router';
+import 'normalize.css/normalize.css';
 import Layout from './Layout';
 import GMap from './GMap';
 import GMapLayers from './GMapLayers';
 import GMapOptim from './GMapOptim';
 import GMapResizable from './GMapResizable';
-
-import 'normalize.css/normalize.css';
 import './Main.sass';
 
 const mountNode = document.getElementById('app');
@@ -22,7 +22,6 @@ render(
       <Route markersCount={20} path="resizable" component={GMapResizable} />
       <IndexRoute markersCount={20} component={GMap} />
     </Route>
-  </Router>
-  ,
+  </Router>,
   mountNode
 );
