@@ -6,31 +6,30 @@ import pure from 'recompose/pure';
 import { Motion, spring } from 'react-motion';
 import clusterMarkerStyles from './ClusterMarker.sass';
 
-export const clusterMarker = ({
-  styles, text, hovered, $hover,
-  defaultMotionStyle, motionStyle,
-}) => (
-  <Motion
-    defaultStyle={defaultMotionStyle}
-    style={motionStyle}
-  >
+export const clusterMarker = (
   {
-    ({ scale }) => (
+    styles,
+    text,
+    hovered,
+    $hover,
+    defaultMotionStyle,
+    motionStyle,
+  }
+) => (
+  <Motion defaultStyle={defaultMotionStyle} style={motionStyle}>
+    {({ scale }) => (
       <div
         className={styles.marker}
         style={{
           transform: `translate3D(0,0,0) scale(${scale}, ${scale})`,
-          zIndex: (hovered || $hover) ? 1 : 0,
+          zIndex: hovered || $hover ? 1 : 0,
         }}
       >
-        <div
-          className={styles.text}
-        >
+        <div className={styles.text}>
           {text}
         </div>
       </div>
-    )
-  }
+    )}
   </Motion>
 );
 
@@ -49,29 +48,33 @@ export const clusterMarkerHOC = compose(
   // pure optimization can cause some effects you don't want,
   // don't use it in development for markers
   pure,
-  withPropsOnChange(
-    ['initialScale'],
-    ({ initialScale, defaultScale, $prerender }) => ({
-      initialScale,
-      defaultMotionStyle: { scale: $prerender ? defaultScale : initialScale },
-    })
-  ),
-  withPropsOnChange(
-    ['hovered', '$hover'],
-    ({
-      hovered, $hover, hoveredScale, defaultScale,
-      stiffness, damping, precision,
-    }) => ({
-      $hover,
-      hovered,
-      motionStyle: {
-        scale: spring(
-          (hovered || $hover) ? hoveredScale : defaultScale,
-          { stiffness, damping, precision }
-        ),
-      },
-    })
-  )
+  withPropsOnChange(['initialScale'], ({
+    initialScale,
+    defaultScale,
+    $prerender,
+  }) => ({
+    initialScale,
+    defaultMotionStyle: { scale: $prerender ? defaultScale : initialScale },
+  })),
+  withPropsOnChange(['hovered', '$hover'], ({
+    hovered,
+    $hover,
+    hoveredScale,
+    defaultScale,
+    stiffness,
+    damping,
+    precision,
+  }) => ({
+    $hover,
+    hovered,
+    motionStyle: {
+      scale: spring(hovered || $hover ? hoveredScale : defaultScale, {
+        stiffness,
+        damping,
+        precision,
+      }),
+    },
+  }))
 );
 
 export default clusterMarkerHOC(clusterMarker);
