@@ -3,8 +3,9 @@ import isReferentiallyTransparentFunctionComponent
   from './isReferentiallyTransparentFunctionComponent';
 
 const createFactory = type => {
-  const isReferentiallyTransparent =
-    isReferentiallyTransparentFunctionComponent(type);
+  const isReferentiallyTransparent = isReferentiallyTransparentFunctionComponent(
+    type
+  );
   return (p, c) =>
     createEagerElementUtil(false, isReferentiallyTransparent, type, p, c);
 };
