@@ -3,7 +3,7 @@ const omit = (obj, keys) => {
   const { ...rest } = obj;
   for (let i = 0; i < keys.length; i++) {
     const key = keys[i];
-    if (rest.hasOwnProperty(key)) {
+    if (key in rest) {
       delete rest[key];
     }
   }
