@@ -4,7 +4,7 @@ const pick = (obj, keys) => {
   const result = {};
   for (let i = 0; i < keys.length; i++) {
     const key = keys[i];
-    if (obj.hasOwnProperty(key)) {
+    if (key in obj) {
       result[key] = obj[key];
     }
   }
