@@ -1,2 +1,3 @@
-import GoogleMap from './google_map.js';
+import GoogleMap from './google_map';
+
 module.exports = GoogleMap;
