@@ -10,38 +10,47 @@
 **/
 
 // Reliable `window` and `document` detection
-var canUseDOM = !!(
-    (typeof window !== 'undefined' &&
-    window.document && window.document.createElement)
-);
+var canUseDOM = !!(typeof window !== 'undefined' &&
+  window.document &&
+  window.document.createElement);
 
 // Check `document` and `window` in case of server-side rendering
-var _window
+var _window;
 if (canUseDOM) {
-  _window = window
+  _window = window;
 } else if (typeof self !== 'undefined') {
-  _window = self
+  _window = self;
 } else {
-  _window = this
+  _window = this;
 }
 
 var attachEvent = typeof document !== 'undefined' && document.attachEvent;
 var stylesCreated = false;
 
 if (canUseDOM && !attachEvent) {
-  var requestFrame = (function(){
-    var raf = _window.requestAnimationFrame || _window.mozRequestAnimationFrame || _window.webkitRequestAnimationFrame ||
-              function(fn){ return _window.setTimeout(fn, 20); };
-    return function(fn){ return raf(fn); };
+  var requestFrame = (function() {
+    var raf = _window.requestAnimationFrame ||
+      _window.mozRequestAnimationFrame ||
+      _window.webkitRequestAnimationFrame ||
+      function(fn) {
+        return _window.setTimeout(fn, 20);
+      };
+    return function(fn) {
+      return raf(fn);
+    };
   })();
 
-  var cancelFrame = (function(){
-    var cancel = _window.cancelAnimationFrame || _window.mozCancelAnimationFrame || _window.webkitCancelAnimationFrame ||
-                 _window.clearTimeout;
-    return function(id){ return cancel(id); };
+  var cancelFrame = (function() {
+    var cancel = _window.cancelAnimationFrame ||
+      _window.mozCancelAnimationFrame ||
+      _window.webkitCancelAnimationFrame ||
+      _window.clearTimeout;
+    return function(id) {
+      return cancel(id);
+    };
   })();
 
-  var resetTriggers = function(element){
+  var resetTriggers = function(element) {
     var triggers = element.__resizeTriggers__,
       expand = triggers.firstElementChild,
       contract = triggers.lastElementChild,
@@ -54,20 +63,20 @@ if (canUseDOM && !attachEvent) {
     expand.scrollTop = expand.scrollHeight;
   };
 
-  var checkTriggers = function(element){
+  var checkTriggers = function(element) {
     return element.offsetWidth != element.__resizeLast__.width ||
-           element.offsetHeight != element.__resizeLast__.height;
-  }
+      element.offsetHeight != element.__resizeLast__.height;
+  };
 
-  var scrollListener = function(e){
+  var scrollListener = function(e) {
     var element = this;
     resetTriggers(this);
     if (this.__resizeRAF__) cancelFrame(this.__resizeRAF__);
-    this.__resizeRAF__ = requestFrame(function(){
+    this.__resizeRAF__ = requestFrame(function() {
       if (checkTriggers(element)) {
         element.__resizeLast__.width = element.offsetWidth;
         element.__resizeLast__.height = element.offsetHeight;
-        element.__resizeListeners__.forEach(function(fn){
+        element.__resizeListeners__.forEach(function(fn) {
           fn.call(element, e);
         });
       }
@@ -80,20 +89,24 @@ if (canUseDOM && !attachEvent) {
     keyframeprefix = '',
     animationstartevent = 'animationstart',
     domPrefixes = 'Webkit Moz O ms'.split(' '),
-    startEvents = 'webkitAnimationStart animationstart oAnimationStart MSAnimationStart'.split(' '),
-    pfx  = '';
+    startEvents = 'webkitAnimationStart animationstart oAnimationStart MSAnimationStart'.split(
+      ' '
+    ),
+    pfx = '';
 
-  if(canUseDOM) {
+  if (canUseDOM) {
     var elm = document.createElement('fakeelement');
-    if( elm.style.animationName !== undefined ) { animation = true; }
+    if (elm.style.animationName !== undefined) {
+      animation = true;
+    }
 
-    if( animation === false ) {
-      for( var i = 0; i < domPrefixes.length; i++ ) {
-        if( elm.style[ domPrefixes[i] + 'AnimationName' ] !== undefined ) {
-          pfx = domPrefixes[ i ];
+    if (animation === false) {
+      for (var i = 0; i < domPrefixes.length; i++) {
+        if (elm.style[domPrefixes[i] + 'AnimationName'] !== undefined) {
+          pfx = domPrefixes[i];
           animationstring = pfx + 'Animation';
           keyframeprefix = '-' + pfx.toLowerCase() + '-';
-          animationstartevent = startEvents[ i ];
+          animationstartevent = startEvents[i];
           animation = true;
           break;
         }
@@ -102,16 +115,25 @@ if (canUseDOM && !attachEvent) {
   }
 
   var animationName = 'resizeanim';
-  var animationKeyframes = '@' + keyframeprefix + 'keyframes ' + animationName + ' { from { opacity: 0; } to { opacity: 0; } } ';
-  var animationStyle = keyframeprefix + 'animation: 1ms ' + animationName + '; ';
+  var animationKeyframes = '@' +
+    keyframeprefix +
+    'keyframes ' +
+    animationName +
+    ' { from { opacity: 0; } to { opacity: 0; } } ';
+  var animationStyle = keyframeprefix +
+    'animation: 1ms ' +
+    animationName +
+    '; ';
 }
 
 var createStyles = function() {
   if (!stylesCreated) {
     //opacity:0 works around a chrome bug https://code.google.com/p/chromium/issues/detail?id=286360
     var css = (animationKeyframes ? animationKeyframes : '') +
-        '.resize-triggers { ' + (animationStyle ? animationStyle : '') + 'visibility: hidden; opacity: 0; } ' +
-        '.resize-triggers, .resize-triggers > div, .contract-trigger:before { content: \" \"; display: block; position: absolute; top: 0; left: 0; height: 100%; width: 100%; overflow: hidden; } .resize-triggers > div { background: #eee; overflow: auto; } .contract-trigger:before { width: 200%; height: 200%; }',
+      '.resize-triggers { ' +
+      (animationStyle ? animationStyle : '') +
+      'visibility: hidden; opacity: 0; } ' +
+      '.resize-triggers, .resize-triggers > div, .contract-trigger:before { content: " "; display: block; position: absolute; top: 0; left: 0; height: 100%; width: 100%; overflow: hidden; } .resize-triggers > div { background: #eee; overflow: auto; } .contract-trigger:before { width: 200%; height: 200%; }',
       head = document.head || document.getElementsByTagName('head')[0],
       style = document.createElement('style');
 
@@ -125,51 +147,64 @@ var createStyles = function() {
     head.appendChild(style);
     stylesCreated = true;
   }
-}
+};
 
-var addResizeListener = function(element, fn){
+var addResizeListener = function(element, fn) {
   if (element.parentNode === undefined) {
     var tempParentDiv = document.createElement('div');
     element.parentNode = tempParentDiv;
   }
   element = element.parentNode;
-  if (attachEvent) element.attachEvent('onresize', fn);
+  if (attachEvent)
+    element.attachEvent('onresize', fn);
   else {
     if (!element.__resizeTriggers__) {
-      if (getComputedStyle(element).position == 'static') element.style.position = 'relative';
+      if (getComputedStyle(element).position == 'static')
+        element.style.position = 'relative';
       createStyles();
       element.__resizeLast__ = {};
       element.__resizeListeners__ = [];
-      (element.__resizeTriggers__ = document.createElement('div')).className = 'resize-triggers';
+      (element.__resizeTriggers__ = document.createElement(
+        'div'
+      )).className = 'resize-triggers';
       element.__resizeTriggers__.innerHTML = '<div class="expand-trigger"><div></div></div>' +
-                                          '<div class="contract-trigger"></div>';
+        '<div class="contract-trigger"></div>';
       element.appendChild(element.__resizeTriggers__);
       resetTriggers(element);
       element.addEventListener('scroll', scrollListener, true);
 
       /* Listen for a css animation to detect element display/re-attach */
-      animationstartevent && element.__resizeTriggers__.addEventListener(animationstartevent, function(e) {
-        if(e.animationName == animationName)
-          resetTriggers(element);
-      });
+      animationstartevent &&
+        element.__resizeTriggers__.addEventListener(
+          animationstartevent,
+          function(e) {
+            if (e.animationName == animationName) resetTriggers(element);
+          }
+        );
     }
     element.__resizeListeners__.push(fn);
   }
 };
 
-var removeResizeListener = function(element, fn){
+var removeResizeListener = function(element, fn) {
   element = element.parentNode;
-  if (attachEvent) element.detachEvent('onresize', fn);
+  if (attachEvent)
+    element.detachEvent('onresize', fn);
   else {
-    element.__resizeListeners__.splice(element.__resizeListeners__.indexOf(fn), 1);
+    element.__resizeListeners__.splice(
+      element.__resizeListeners__.indexOf(fn),
+      1
+    );
     if (!element.__resizeListeners__.length) {
-        element.removeEventListener('scroll', scrollListener);
-        element.__resizeTriggers__ = !element.removeChild(element.__resizeTriggers__);
+      element.removeEventListener('scroll', scrollListener);
+      element.__resizeTriggers__ = !element.removeChild(
+        element.__resizeTriggers__
+      );
     }
   }
-}
+};
 
 module.exports = {
-  addResizeListener : addResizeListener,
-  removeResizeListener : removeResizeListener
+  addResizeListener: addResizeListener,
+  removeResizeListener: removeResizeListener,
 };
