@@ -1,11 +1,10 @@
-import LatLng from './lib_geo/lat_lng.js';
 import Point from 'point-geometry';
-import Transform from './lib_geo/transform.js';
-
+import LatLng from './lib_geo/lat_lng';
+import Transform from './lib_geo/transform';
 
 export default class Geo {
-
-  constructor(tileSize) { // left_top view пользует гугл
+  constructor(tileSize) {
+    // left_top view пользует гугл
     // super();
     this.hasSize_ = false;
     this.hasView_ = false;
@@ -51,7 +50,8 @@ export default class Geo {
   project(ptLatLng, viewFromLeftTop) {
     if (viewFromLeftTop) {
       const pt = this.transform_.locationPoint(LatLng.convert(ptLatLng));
-      pt.x -= this.transform_.worldSize * Math.round(pt.x / this.transform_.worldSize);
+      pt.x -= this.transform_.worldSize *
+        Math.round(pt.x / this.transform_.worldSize);
 
       pt.x += this.transform_.width / 2;
       pt.y += this.transform_.height / 2;
@@ -81,12 +81,14 @@ export default class Geo {
   }
 
   getBounds(margins, roundFactor) {
-    const bndT = margins && margins[0] || 0;
-    const bndR = margins && margins[1] || 0;
-    const bndB = margins && margins[2] || 0;
-    const bndL = margins && margins[3] || 0;
+    const bndT = (margins && margins[0]) || 0;
+    const bndR = (margins && margins[1]) || 0;
+    const bndB = (margins && margins[2]) || 0;
+    const bndL = (margins && margins[3]) || 0;
 
-    if (this.getWidth() - bndR - bndL > 0 && this.getHeight() - bndT - bndB > 0) {
+    if (
+      this.getWidth() - bndR - bndL > 0 && this.getHeight() - bndT - bndB > 0
+    ) {
       const topLeftCorner = this.unproject({
         x: bndL - this.getWidth() / 2,
         y: bndT - this.getHeight() / 2,
@@ -97,10 +99,14 @@ export default class Geo {
       });
 
       let res = [
-        topLeftCorner.lat, topLeftCorner.lng, // NW
-        bottomRightCorner.lat, bottomRightCorner.lng, // SE
-        bottomRightCorner.lat, topLeftCorner.lng, // SW
-        topLeftCorner.lat, bottomRightCorner.lng, // NE
+        topLeftCorner.lat,
+        topLeftCorner.lng, // NW
+        bottomRightCorner.lat,
+        bottomRightCorner.lng, // SE
+        bottomRightCorner.lat,
+        topLeftCorner.lng, // SW
+        topLeftCorner.lat,
+        bottomRightCorner.lng, // NE
       ];
 
       if (roundFactor) {
