@@ -1,5 +1,5 @@
 // source taken from https://github.com/rackt/redux/blob/master/src/utils/isPlainObject.js
-const fnToString = (fn) => Function.prototype.toString.call(fn);
+const fnToString = fn => Function.prototype.toString.call(fn);
 
 /**
  * @param {any} obj The object to inspect.
@@ -20,7 +20,7 @@ export default function isPlainObject(obj) {
 
   const constructor = proto.constructor;
 
-  return typeof constructor === 'function'
-    && constructor instanceof constructor
-    && fnToString(constructor) === fnToString(Object);
+  return typeof constructor === 'function' &&
+    constructor instanceof constructor &&
+    fnToString(constructor) === fnToString(Object);
 }
